Commit before re-implementing a lot of things
I ran into some problems with implicit state and the control flow is a
bit confusing to me.
The code up until now was heavily inspired by the GLES3 code.

I'll try to do it from a more bottom-up approach, so I'll re-implement
everything so far, but differently.

This is just a "save point", so nothing too interesting to see here.